Add cached entry to harvester.
